Refactor: Centralize calculateDistance and clarify findActiveClusters duplication
- I moved `calculateDistance` to `common/mathUtils.js`.
- I created a synchronized copy `functions/utils/mathUtils.js` for backend use.
- I updated all imports of `calculateDistance` in both your frontend and backend code to use these new locations.
- I adjusted unit tests (mocks and imports) to reflect these changes.
- I corrected `calculateDistance` imports in `RegionalSeismicityChart.jsx` and `EarthquakeDetailView.jsx`.
- I updated comments in both frontend and backend versions of `findActiveClusters` to:
    - Clarify their relationship and specific differences (e.g., backend duplicate check, logging).
    - Indicate that `calculateDistance` is now sourced from the centralized/synced utility files.
- I marked Task 2 in `improvement_tasks.txt` as complete.

The build is successful and tests related to these changes are passing. Pre-existing unrelated test failures were not addressed by this change.
