+++ src/utils/utils.js
@@ -6,29 +6,7 @@
  * including calculations, formatting, validation, and color utilities.
  */
 
-// NOTE: This function is duplicated in functions/api/calculate-clusters.js
 // Any algorithmic changes should be synchronized.
-/**
- * Calculates the distance between two geographical coordinates using the Haversine formula.
- * @param {number} lat1 Latitude of the first point.
- * @param {number} lon1 Longitude of the first point.
- * @param {number} lat2 Latitude of the second point.
- * @param {number} lon2 Longitude of the second point.
- * @returns {number} Distance in kilometers.
- */
-export function calculateDistance(lat1, lon1, lat2, lon2) {
-    const R = 6371; // Radius of the Earth in kilometers
-    const dLat = (lat2 - lat1) * Math.PI / 180;
-    const dLon = (lon2 - lon1) * Math.PI / 180;
-    const a =
-        Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-        Math.cos(lat1 * Math.PI / 180) * Math.cos(lat2 * Math.PI / 180) *
-        Math.sin(dLon / 2) * Math.sin(dLon / 2);
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-    const distance = R * c;
-    return distance;
-}
-
 /**
  * Returns a hex color code based on earthquake magnitude.
  * @param {number | null | undefined} magnitude - The earthquake magnitude.
